Remove ioctl for changing BUSY pin
BUSY pin mapping can be changed via ioctl. However, this implementation is
not fully tested and seems not to work. Moreover, BUSY pin mapping
is to be done static via device tree. There is no need to change it
during runtime of the module.

This removes the code for changing BUSY pin within ioctl.