Interrupt based BUSY line implementation
This removes busy-polling of BUSY line and uses IRQ based handling
instead. While the former is a more performant solution (in a sense
that it minimizes delays of back-to-back transfers) is the latter
more cooperate solution (task sleep instead of busy polling).
The measured latency on a raspi3 running at 1.2GHz between
two back to back transfers is increasing from 2us to 10us.

Signed-off-by: Christian Eisendle <christian@eisendle.net>